earth_tester should only reset test db if you request it

+++ bin/earth_tester.rb
@@ -32,7 +32,7 @@ when /postgr/i
 when /sqlite/i
   ActiveRecord::Base.establish_connection(:adapter => "sqlite3", :database => ":memory:")
 else
-  system %{mysql -u root -ppassword -e "DROP DATABASE test_earth"}
+  system %{mysql -u root -ppassword -e "DROP DATABASE test_earth"} if ENV['RESET_DB'] == 'true'
   system %{mysql -u root -ppassword -e "CREATE DATABASE test_earth CHARSET utf8"}
   ActiveRecord::Base.establish_connection(
     'adapter' => (RUBY_PLATFORM == 'java' ? 'mysql' : 'mysql2'),
